+++ fuzz/fuzz_targets/codec_byte_string_read.rs
@@ -1,0 +1,7 @@
+#![no_main]
+
+use libfuzzer_sys::fuzz_target;
+
+fuzz_target!(|data: &[u8]| {
+    let _ = sagnir_codec::read_byte_string(data, 4096);
+});